Resolved merge conflicts to I2C, SPI, NRF functions
