+++ .gitignore
@@ -1,3 +1,4 @@
+/PSLab_Original/nbproject/private/
 /PSLab_Original/build/default/
 /PSLab_Original/dist/default/
 /MPLABProject/nbproject/private/
