+++ PSLab_Original/nbproject/configurations.xml
@@ -7,6 +7,10 @@
       <itemPath>COMMANDS.h</itemPath>
       <itemPath>functions.h</itemPath>
       <itemPath>PSLAB_UART.h</itemPath>
+      <itemPath>PSLAB_I2C.h</itemPath>
+      <itemPath>Common_Functions.h</itemPath>
+      <itemPath>PSLAB_NRF.h</itemPath>
+      <itemPath>PSLAB_SPI.h</itemPath>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
                    displayName="Linker Files"
@@ -17,6 +21,10 @@
                    projectFiles="true">
       <itemPath>proto2_main.c</itemPath>
       <itemPath>PSLAB_UART.c</itemPath>
+      <itemPath>PSLAB_I2C.c</itemPath>
+      <itemPath>Common_Functions.c</itemPath>
+      <itemPath>PSLAB_NRF.c</itemPath>
+      <itemPath>PSLAB_SPI.c</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
