+++ PSLab_Original/proto2_main.c
@@ -16,7 +16,11 @@
 #include <stdlib.h>
 #include <libpic30.h>
 #include "functions.c"
+#include "Common_Functions.h"
 #include "PSLAB_UART.h"
+#include "PSLAB_I2C.h"
+#include "PSLAB_NRF.h"
+#include "PSLAB_SPI.h"
 //_FICD(ICS_PGD2 & JTAGEN_OFF) //Programmming pins  ..PGED2
